WIP: independent CPU usage in /proc/stat
Containers can see utilization of all available CPUs, even if the CPU
is utilized by other container or by the host. The contents
of `/proc/stat` are shared across the system, except for hiding CPUs
excluded by cpuset. This commit attempts to fix that, but at a cost.
CPU usage is read from cpuacct cgroup, but that accounts only for
`user` and `system` fields from `/proc/stat`. Idle time can be
calculated, but other fields cannot, thus are set to 0.